fix: do not drop commits without known type
Projects may have different conventions that those mandated by the
conventional-commits-parser[1] package. For example, it is customary
to also prefix commits depending on the area of change such as

src: fix foobar
tests: Fix another test

All these commits, when they are parsed, they end up with type == "src"
or type == "tests" and the package was dropping them silently because
they did not have a known type. However, we shouldn't do that but we
should let projects be flexible on how they use their commit messages
and simply list them along with those of type == null if they do not have a
known type.

[1] https://www.npmjs.com/package/conventional-commits-parser