feat: allow to filter out certain release tags
As we are comparing git tags to determine which one is right before the
one we are releasing right now, it may be useful to filter out some of
them and only get a git changelog from the tags we really care about.

For example:

v1.0.0
v1.0.1-beta
v1.0.1-rc
v1.0.1

Right now, when we tag and release v1.0.1, then the generated changelog
will contain commits that happend between v1.0.1-rc and v1.0.1. However,
this is not always desirable and it may make more sense to have a
changelog betwwen v1.0.0 and v1.0.1. As such, lets provide this
information as regex

Fixes: #25