feat(automatic-releases): export the URL for uploading assets
Sometimes it may be useful to add additional assets to a created release
afterwards (e.g. build multiple binaries for every supported platform
and attach all of them to the same release). To make it possible, we
just need to export the release upload URL as an output of the
automatic-releases action.